Rename & move gRPC handler and marshaler
